[DFT] DFT Regeneration and test for mcp_pmsm_foc_smo_sam_e54_mclv2_sam_e54_pim MH3-81567

+++ apps/pmsm_foc_smo_sam_e54/firmware/src/config/mchv3_sam_e54_pim/QSpin/Field_Oriented_Control/mc_pwm.c
@@ -126,7 +126,7 @@ void  mcPwmI_PulseWidthModulationInit( tmcPwm_Parameters_s * const pParameters )
     pState->maxPeriodCount = pState->pwmPeriodCount - pState->minPeriodCount;
     pState->maxModIndex = pParameters->maxModIndex;
 
-    allowedRange = (float32_t)( pState->maxPeriodCount - pState->minPeriodCount);
+    allowedRange = (float32_t)( (float32_t)pState->maxPeriodCount - (float32_t)pState->minPeriodCount);
     pState->uBusFactor = (float32_t)( allowedRange / (float32_t)pState->pwmPeriodCount );
 }
